+++ marketplace-data/src/main/java/com/marketplace/data/banner/BannerRepo.java
@@ -7,12 +7,12 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface BannerRepo extends JpaRepository<BannerEntity, Integer> {
+public interface BannerRepo extends JpaRepository<BannerEntity, Long> {
 
-    <T> Optional<T> getBannerById(int id, Class<T> type);
+    <T> Optional<T> getBannerById(long id, Class<T> type);
     
     @Modifying
 	@Query("UPDATE Banner b SET b.image = :image WHERE b.id = :id")
-	void updateImage(@Param("id") int id, @Param("image") String image);
+	void updateImage(@Param("id") long id, @Param("image") String image);
 
 }
